init unAvailable func in podlifecycle and update ruleset sample (#50)
* init unAvailable func and update ruleset sample

* fix ut

+++ pkg/controllers/ruleset/ruleset_controller_test.go
@@ -125,7 +125,7 @@ const (
 
 func initRulesetManager() {
 
-	register.UnAvailableFuncList = append(register.UnAvailableFuncList, func(pod *corev1.Pod) (bool, *int32) {
+	register.UnAvailableFuncList = append(register.UnAvailableFuncList, func(pod *corev1.Pod) (bool, *int64) {
 		if pod.GetLabels() == nil {
 			return false, nil
 		}
